refactor cart section
